ベーステストをテストのルートに移動
