--- Assets/Editor/TEST.MyselfTest.cs.meta
@@ -1,13 +1,0 @@
-fileFormatVersion: 2
-guid: d7333f0482c53744a944abb8ab77e9d7
-timeCreated: 1513340559
-licenseType: Free
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
